backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
google-generativeai>=0.3.0
httpx>=0.25.1
python-dotenv>=1.0.0
pillow>=10.0.0

